ip_grib2 branch:
Remove unused argument "li" from routines polatev4 and
polates4.  

Routines ipxwafs, ipxwafs2, ipxwafs3 - initialize
MISSING to "-1", which is all bits on for
2's complement machines.  Was being set to
b'1111s' which the gnu Fortran compiler did not like. 

Several routines - variable ISCALE, which is the
grib 2 scale factor for lat/lon, is an integer.
It was being initialzed using a real value which
the gnu compiler did not like.

Ticket #81.