ip_grib2 branch:
Remove the 'gausslat' routine as it is only a wrapper
around the call to sp library routine 'splat'.
Previous searches showed it was only used by one
OPS program.  Since this grib2 version will be
a major release with numerous interface changes,
this is a logical place to remove 'gausslat'.

Ticket #81.