Re-factor library to use a generic grib_descriptor class
Create an ip_grid_descriptor class that represents a generic grid descriptor that can be created from a grib1 (kgds) or grib2 template array (igdtmpl).

Then those are passed around to subroutines instead of the grib arrays themselves.
This makes the functions more generic and paves the way to combine the ip and ip2 libraries by combining shared code and handling the grib specific code in a single layer

Scalar tests pass

Vectors still need to be implemented